requirements.txt
requests
python-dotenv
deepdiff